values: explicitly mention works council members in the context of DIB
At GitLab we have multiple works councils in different countries by now.
Works councils represent the interests of the employees they have been
elected by in the respective country. As such, their core duties are,
amongst others:

  - To monitor whether the employer observes laws as well as collective
    agreements.

  - To ensure that there is no discrimination against employees.

  - To protect employees in case they are being treated unfairly.

  - To negotiate agreements with the employer as required to ensure the
    wellbeing of employees.

As such, works councils work on eye level with executives and are the
direct contact points for employees in case they have any concerns.

Members of works councils are elected by their employees and have thus
demonstrated that they are deemed trustworthy by their voters to handle
both confidential and highly personal matters. Furthermore, the position
is done voluntarily without any extra compensation.

This commit explicitly includes works council members into our table of
competencies for our Diversion, Inclusion & Belonging value. The intent
is to encourage people to become active in works councils. Furthermore,
it demonstrates that GitLab is supportive of works councils and that it
appreciates the role that they are playing in the company. This is quite
an important aspect given that many companies tend to treat works
councils as enemies rather than as friends, and consequently employees
fear retaliation when becoming active in one. But we found that not to
be the case at GitLab, so by explicitly highlighting the works council
work as achievement helps to demonstrate the spirit of mutual trust.

As mentioned before, works councils work on eye level with executives
and hold them accountable. Furthermore, they implement measures for and
are the direct contact point of the employees of their entity. This
matches the description that we already have in place for job grade 12,
so add works councils there.