rate-limiting: Cloudflare Security Events and SSH Traffic
Adds more details for investigating with Cloudflare, notes that there
isn't much (if any) troubleshooting guidance for SSH traffic, and points
to runbooks where necessary as SREs will be required to do those
investigations.

It also references the HAProxy runbook, as although this guide is
specifically focused on gitlab.com, it makes sense to point to why and
when you might need to look into the HAProxy logs.