rate-limiting: Add note about avoiding rate limits
We added a section within the GitLab docs around avoiding being rate
limiting as a result of authentication ban, prompted by the expiry of
lifetime tokens in May 2024.

This commit takes some of that guidance and adds it in a more generic
area of our handbook, because that guidance was only present under a
specific use case but would be valuable for all customers trying to
minimise the impact of rate limits.

Related: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/152396