rate-limiting: Add link to video walkthrough
A walkthrough of the Cloudflare Dashboard was done to accompany this
guidance, so show a few examples of navigating the dashboard, and how to
use the filters for each of the views.

From this, it was found that those with Cloudflare Analytics access
aren't able to view the names of the rules that blocked a request, but
they _can_ at least see that requests have been blocked by rate limiting
for example. So that's now called out.

Additionally: added a link to the RackAttack Troubleshooting guide in
the GitLab docs.