Fixes the reference for Infra leader
We no longer have a single schedule for the infra leadership rotation.
This changed after we switched from one schedule to a schedule per
person, so we could page people simultaneously.

This changes the schedule link to the section on infra leadership, which
has instructions on how to page them.