Use latest buy subscription link for developer license setup
I was following these instructions and found that this link to
buy_subscription.md was linking to an old SHA. This then led
me to follow the subsequent link to instructions for Duo licenses
using Zuora. This didn't work for me but after getting the latest version
of these instructions I was now able to purchases the Duo Add-On
without logging in to Zuora at all.