Add Redis as data that is not in scope for Backup and Restore
For https://gitlab.com/gitlab-com/gl-infra/data-access/durability/team/-/issues/112
We are proposing that we stop storing GCP disk snapshots for Redis as it
they are not part of our backup and recovery plan for GitLab.com.

This matches what our backup/restore policy for Dedicated, and also
matches what we think we can validate as part of recovery.

- We have very low confidence in restoring from snapshot.
- If we were to restore we would have consistency issues with the
  database since snapshots are on a fixed schedule.