Update to eligibility requirements for Incident Managers
In [this
discussions](https://docs.google.com/document/d/1RcahFw7TGpTPLpKBK45GADeqcancaq2ivXa_dE83T5E/edit?tab=t.0#heading=h.tdq3vzoa2l79)
(internal) we discussed that soon we will changing dev-oncall so that it
is not an "on-call" rotation. Therefore, we can remove references to it
for IM eligibility.