Add security product configuration section
We (application security testing stage) recently discussed how our
configuration direction works in conjuction with our convention over
configuration principle.  The consensus was that our feature set is more
likely to require configuration in comparison to other features due to
the product domain. This subsection clarifies that this is the case, but
makes an effort highlight that while we should still view this as an
absolute last resort.