company: Allow the use of "GitLabber"
In the past we only referred to people working at GitLab as "team
members". This is a bit confusing, given that "team" refers to a small
group. We didn't use "GitLabber" because we considered that to be
applicable to the whole GitLab community. But it actually never gained
any traction.

Allow to use of "GitLabber" to refer to a "GitLab team member". Using it
is very similar to:

* "Googler": people working at Google
* "Metamates": people working at Meta
* "Figmates": people working at Figma

The word "GitLabber" is a lot shorter than "GitLab team member" and
implies a greater sense of camaraderie.