Add an extra con to highlight additional platform constraints
This change highlights the additional requirement that a centralized
configuration would incur on platform implementors relying on
Cloudflare functionality.

This comes from Runway specifically where tenants would need to make
changes outside of the currently central config to enable Cloudflare
functionality. This would be greatly detrimental to the user experience.