Update phase 5 exit criteria
What
---
Mention that unique indexes need to be audited

Why
---
Claiming of resources means that anything that needs to be cluster-wide
unique needs to be claimed. This was initially thought that it means any
UNIQUE index in the database needs to be claimed, but this is not true
since those UNIQUE indexes can/should be scoped to an organization.