Start document of cluster wide unique resources for organizations
What
---
Start a table that will be used to document the cluster-wide resources
even when we have organization isolation.

Why
---
When a customer is using organization they will rightfully assume that
everything is isolated but we have special cases where things have to be
cluster-wide unique to keep backward compatibility.