Adds a release glossary
Introduces a release glossary on the releases page. With the
early-branch and maintenance efforts, it is important to ensure a
terminology is defined across GitLab releases. This commit adds a
seciton on the releases/_index.md page with a release glossary.

It also updates the backports section to remove the old 'security
release' process