fix: show all categories on the categories lookup page
This makes sure that we use stages.yml as the SSoT of feature category
ownership.

categories.yml changes as the product direction changes. This means
that when features get divested, they would no longer be mentioned in
categories.yml, but this does not mean the features are no longer used
by our users or that the code no longer exists. When this happens, we
can move the feature category from group.categories to
group.maintained_categories.

Having it like this means that we can remove the feature category from
categories.yml indicating this is no longer a priority from a product
perspective, while still communicating which group owns that code.

Related to gitlab-org/gitlab#549006 (comment 2622703178)